Move EventAdapter and EventDateDateComparatorShould to eventseries package
