+++ app/src/main/java/org/gdg/frisbee/android/eventseries/EventListFragment.java
@@ -29,7 +29,6 @@ import android.widget.ListView;
 
 import org.gdg.frisbee.android.Const;
 import org.gdg.frisbee.android.R;
-import org.gdg.frisbee.android.adapter.EventAdapter;
 import org.gdg.frisbee.android.api.model.SimpleEvent;
 import org.gdg.frisbee.android.event.EventActivity;
 import org.gdg.frisbee.android.common.GdgListFragment;
