+++ app/src/main/java/org/gdg/frisbee/android/eventseries/GdgEventListFragment.java
@@ -4,7 +4,6 @@ import android.os.Bundle;
 
 import org.gdg.frisbee.android.Const;
 import org.gdg.frisbee.android.R;
-import org.gdg.frisbee.android.adapter.EventAdapter;
 import org.gdg.frisbee.android.api.model.Event;
 import org.gdg.frisbee.android.app.App;
 import org.gdg.frisbee.android.cache.ModelCache;
