+++ app/src/test/java/org/gdg/frisbee/android/eventseries/EventDateComparatorShould.java
@@ -1,9 +1,7 @@
-package org.gdg.frisbee.android.utils;
+package org.gdg.frisbee.android.eventseries;
 
-import org.gdg.frisbee.android.adapter.EventAdapter;
 import org.gdg.frisbee.android.api.model.Event;
 import org.gdg.frisbee.android.api.model.SimpleEvent;
-import org.gdg.frisbee.android.eventseries.EventDateComparator;
 import org.joda.time.DateTime;
 import org.junit.Before;
 import org.junit.Test;
